Disable ghost touches temporarily
The touch sensor sometimes inputs ghost touches and stays on for a long
time. This can be due to atmospherical conditions or insects etc.. Now
the touch sensor will be disabled when these events occur and it will
be re-enabled if the condition is removed. Control via MQTT is always
available.